feat(DTFS2-7171): refactor about-facility validation
