+++ gef-ui/server/routes/about-facility.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const { aboutFacility, validateAboutFacility } = require('../controllers/about-facility');
+const { aboutFacility, validateAndUpdateAboutFacility } = require('../controllers/about-facility');
 const { validateRole, validateToken, validateBank } = require('../middleware');
 const { MAKER } = require('../constants/roles');
 
@@ -9,7 +9,7 @@ router.get('/application-details/:dealId/facilities/:facilityId/about-facility',
   aboutFacility(req, res),
 );
 router.post('/application-details/:dealId/facilities/:facilityId/about-facility', [validateToken, validateBank, validateRole({ role: [MAKER] })], (req, res) =>
-  validateAboutFacility(req, res),
+  validateAndUpdateAboutFacility(req, res),
 );
 
 module.exports = router;
